Setting the background programs in a crontab

+++ playbooks/files/bin/system-builder.sh
@@ -1,13 +1,27 @@
 #!/bin/sh -eu
 
-LOCAL_REV=${1}
-REMOTE_REV=${2}
-URL=${3}
-SRC_PATH=${4:-/usr/src}
+get_svn_info()
+{
+	local path=${1}
+	local param=${2}
+	svnlite info "${path}" | grep -E "^${param}:" | \
+		sed "s,^${param}: ,,"
+}
+
+SRC_PATH=${1:-/usr/src}
+LOCAL_REV=$(get_svn_info "${SRC_PATH}" "Last Changed Rev")
+URL=$(get_svn_info "${SRC_PATH}" "URL")
+REMOTE_REV=$(get_svn_info "${URL}" "Last Changed Rev")
 
 DIFF_FILE=$(mktemp /tmp/system-builder.XXXXXX)
 svn diff --summarize --revision "${LOCAL_REV}:${REMOTE_REV}" \
 	"${URL}" > "${DIFF_FILE}"
+
+echo "${LOCAL_REV}" > /var/log/system-builder
+if [ "$(find /var/run -name 'system-builder.*')" ] ; then
+	exit 0
+fi
+
 svn up "${SRC_PATH}"
 
 REGEX="^[^[:space:]]+[[:space:]]+${URL}/sys/"
@@ -20,10 +34,14 @@ if grep -qvE "${REGEX}" "${DIFF_FILE}" ; then
 	TARGETS="${TARGETS} buildworld"
 fi
 
+if [ -z "${TARGETS}" ] ; then
+	exit 0
+fi
+
 for target in ${TARGETS} ; do
-	touch "/var/run/system-builder.${target}"
+	echo "${LOCAL_REV} -> ${REMOTE_REV}" > "/var/run/system-builder.${target}"
 	cd "${SRC_PATH}" && make "${target}"
 	rm "/var/run/system-builder.${target}"
 done
 
-touch "/var/run/system-builder.ready"
+echo "${LOCAL_REV} -> ${REMOTE_REV}" > "/var/run/system-builder.ready"
